% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udpipe_parse.R
\name{as.data.frame.udpipe_connlu}
\alias{as.data.frame.udpipe_connlu}
\title{Convert the result of udpipe_annotate to a tidy data frame}
\usage{
\method{as.data.frame}{udpipe_connlu}(x, ...)
}
\arguments{
\item{x}{an object of class \code{udpipe_connlu} as returned by \code{\link{udpipe_annotate}}}

\item{...}{currently not used}
}
\value{
a data.frame with columns 
doc_id, paragraph_id, sentence_id, sentence_text, 
id, form, lemma, upostag, xpostag, feats, head, deprel, deps, misc)
}
\description{
Convert the result of udpipe_annotate to a tidy data frame
}
\examples{
\dontrun{
## Load the model
f <- file.path(getwd(), "dev/udpipe-ud-2.0-170801/dutch-ud-2.0-170801.udpipe")
ud_dutch <- udpipe_load_model(f)

## Tokenise, Tag and Dependency Parsing Annotation. Output is in CONLL-U format.
txt <- c("Dus. Godvermehoeren met pus in alle puisten, 
  zei die schele van Van Bukburg en hij had nog gelijk ook. 
  Er was toen dat liedje van tietenkonttieten kont tieten kontkontkont, 
  maar dat hoefden we geenseens niet te zingen. 
  Je kunt zeggen wat je wil van al die gesluierde poezenpas maar d'r kwam wel 
  een vleeswarenwinkel onder te voorschijn van heb je me daar nou.
  
  En zo gaat het maar door.",
  "Wat die ransaap van een academici nou weer in z'n botte pan heb gehaald mag 
  Joost in m'n schoen gooien, maar feit staat boven water dat het een gore 
  vieze vuile ransaap is.")
x <- udpipe_annotate(ud_dutch, x = txt)
as.data.frame(x)
}
}
\seealso{
\code{\link{udpipe_annotate}}
}
