% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{brussels_listings}
\alias{brussels_listings}
\title{Brussels AirBnB address locations available at www.insideairbnb.com}
\source{
https://insideairbnb.com/brussels: information of 2015-10-03
}
\description{
Brussels AirBnB address locations available at www.insideairbnb.com
More information: https://insideairbnb.com/get-the-data.html \cr
Data has been converted from UTF-8 to ASCII as in \code{iconv(x, from = "UTF-8", to = "ASCII//TRANSLIT")} in order
to be able to comply to CRAN policies.
}
\examples{
data(brussels_listings)
head(brussels_listings)
}
\seealso{
\code{\link{brussels_reviews}}, \code{\link{brussels_reviews_anno}}
}
