% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{brussels_reviews}
\alias{brussels_reviews}
\title{Reviews of AirBnB customers on Brussels address locations available at www.insideairbnb.com}
\source{
\url{http://insideairbnb.com/brussels}: information of 2015-10-03
}
\description{
Reviews of AirBnB customers on Brussels address locations available at www.insideairbnb.com
More information: http://insideairbnb.com/get-the-data.html.
The data contains 500 reviews in Spanish, 500 reviews in French and 500 reviews in Dutch.\cr
The data frame contains the field id (unique), listing_id which corresponds to the listing_id of
the \code{\link{brussels_listings}} dataset and text fields feedback and language (identified with package cld2) \cr
Data has been converted from UTF-8 to ASCII as in \code{iconv(x, from = "UTF-8", to = "ASCII//TRANSLIT")} in order
to be able to comply to CRAN policies.
}
\examples{
data(brussels_reviews)
str(brussels_reviews)
head(brussels_reviews)
}
\seealso{
\code{\link{brussels_listings}}, \code{\link{brussels_reviews_anno}}
}
