% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{brussels_reviews_w2v_embeddings_lemma_nl}
\alias{brussels_reviews_w2v_embeddings_lemma_nl}
\title{An example matrix of word embeddings}
\description{
An simple 10-dimensional example matrix of word embeddings trained on the Dutch lemma's 
of the dataset \code{\link{brussels_reviews_anno}}
}
\examples{
data(brussels_reviews_w2v_embeddings_lemma_nl)
head(brussels_reviews_w2v_embeddings_lemma_nl)
}
