% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strsplit.data.frame}
\alias{strsplit.data.frame}
\title{Split text alongside a regular expression to obtain a tokenised data frame}
\usage{
strsplit.data.frame(data, term, group,
  split = "[[:space:][:punct:][:digit:]]+")
}
\arguments{
\item{data}{a data.frame or data.table}

\item{term}{a column name from \code{data} which you want to split into tokens}

\item{group}{a column name or a vector of column names from \code{data} indicating identifiers of groups. 
Text will be split by group.}

\item{split}{a regular expression indicating how to split the \code{term} column. 
Defaults to splitting by spaces, punctuation symbols or digits.}
}
\value{
a data.frame with the columns from \code{group} and \code{term} where the text in column \code{term}
will be split by the provided regular expression into tokens. The result is a tokenised data frame
containing one row per token.
}
\description{
Split text alongside a regular expression to obtain a tokenised data frame. 
This is the inverse of \code{\link{paste.data.frame}}.
}
\examples{
data(brussels_reviews, package = "udpipe")
x <- strsplit.data.frame(brussels_reviews, term = "feedback", group = "id")
x <- strsplit.data.frame(brussels_reviews, 
                         term = c("feedback"), 
                         group = c("listing_id", "language"))
}
\seealso{
\code{\link{paste.data.frame}}
}
