% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_dependency_utlities.R
\docType{class}
\name{syntaxpatterns-class}
\alias{syntaxpatterns-class}
\alias{syntaxpatterns}
\title{Experimental and undocumented querying of syntax patterns}
\description{
Currently undocumented
}

