% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_dependency_utlities.R
\docType{class}
\name{syntaxrelation-class}
\alias{syntaxrelation-class}
\alias{syntaxrelation}
\alias{|,syntaxrelation,logical-method}
\alias{|,logical,syntaxrelation-method}
\alias{&,syntaxrelation,logical-method}
\alias{&,logical,syntaxrelation-method}
\title{Experimental and undocumented querying of syntax relationships}
\usage{
\S4method{|}{syntaxrelation,logical}(e1, e2)

\S4method{|}{logical,syntaxrelation}(e1, e2)

\S4method{&}{syntaxrelation,logical}(e1, e2)

\S4method{&}{logical,syntaxrelation}(e1, e2)
}
\arguments{
\item{e1}{Currently undocumented}

\item{e2}{Currently undocumented}
}
\description{
Currently undocumented
}

