% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{txt_collapse}
\alias{txt_collapse}
\title{Collapse a character vector while removing missing data.}
\usage{
txt_collapse(x, collapse = " ")
}
\arguments{
\item{x}{a character vector}

\item{collapse}{a character string to be used to collapse the vector. Defaults to a space: ' '.}
}
\value{
a character vector of length 1 with the content of x collapsed using paste
}
\description{
Collapse a character vector while removing missing data.
}
\examples{
txt_collapse(c(NA, "hello", "world", NA))
}
\seealso{
\code{\link{paste}}
}
