% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{txt_overlap}
\alias{txt_overlap}
\title{Get the overlap between 2 vectors}
\usage{
txt_overlap(x, y, ...)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
a vector with elements of \code{x} which are also found in \code{y}
}
\description{
Get the overlap between 2 vectors
}
\examples{
x <- c("a", "b", "c")
y <- c("b", "c", "e", "z")
txt_overlap(x, y)
txt_overlap(y, x)
}
