% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{txt_sample}
\alias{txt_sample}
\title{Boilerplate function to sample one element from a vector.}
\usage{
txt_sample(x, na.exclude = TRUE, n = 1)
}
\arguments{
\item{x}{a vector}

\item{na.exclude}{logical indicating to remove NA values before taking a sample}

\item{n}{integer indicating the number of items to sample from \code{x}}
}
\value{
one element sampled from the vector x
}
\description{
Boilerplate function to sample one element from a vector.
}
\examples{
txt_sample(c(NA, "hello", "world", NA))
}
\seealso{
\code{\link{sample.int}}
}
