% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{txt_show}
\alias{txt_show}
\title{Boilerplate function to cat only 1 element of a character vector.}
\usage{
txt_show(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
invisible
}
\description{
Boilerplate function to cat only 1 element of a character vector.
}
\examples{
txt_show(c("hello \n\n\n world", "world \n\n\n hello"))
}
\seealso{
\code{\link{txt_sample}}
}
