% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udpipe_parse.R
\name{udpipe_annotate}
\alias{udpipe_annotate}
\title{Tokenise, Tag and Dependency Parsing Annotation on text}
\usage{
udpipe_annotate(object, x, doc_id = paste("d", seq_along(x), sep = ""), ...)
}
\arguments{
\item{object}{an object of class \code{udpipe_model} as returned by \code{\link{udpipe_load_model}}}

\item{x}{a character vector in UTF-8 encoding where each element of the character vector 
contains text which you like to tokenize, tag and perform dependency parsing.}

\item{doc_id}{an identifier of a document with the same length as \code{x}.}

\item{...}{currently not used}
}
\value{
a list with 3 elements
\itemize{
 \item{x: }{The \code{x} character vector with text.}
 \item{conllu: }{A character vector of length 1 containing the annotated result of the annotation flow in CONLL-U format.
 This format is explained at \url{http://universaldependencies.org/format.html}}
 \item{error: }{A vector with the same length of \code{x} containing possible errors when annotating \code{x}}
}
}
\description{
Tokenise, Tag and Dependency Parsing Annotation on text
}
\examples{
\dontrun{
## Load the model
f <- file.path(getwd(), "dev/udpipe-ud-2.0-170801/dutch-ud-2.0-170801.udpipe")
ud_dutch <- udpipe_load_model(f)

## Tokenise, Tag and Dependency Parsing Annotation. Output is in CONLL-U format.
txt <- c("Dus. Godvermehoeren met pus in alle puisten, 
  zei die schele van Van Bukburg en hij had nog gelijk ook. 
  Er was toen dat liedje van tietenkonttieten kont tieten kontkontkont, 
  maar dat hoefden we geenseens niet te zingen. 
  Je kunt zeggen wat je wil van al die gesluierde poezenpas maar d'r kwam wel 
  een vleeswarenwinkel onder te voorschijn van heb je me daar nou.
  
  En zo gaat het maar door.",
  "Wat die ransaap van een academici nou weer in z'n botte pan heb gehaald mag 
  Joost in m'n schoen gooien, maar feit staat boven water dat het een gore 
  vieze vuile ransaap is.")
x <- udpipe_annotate(ud_dutch, x = txt)
cat(x$conllu)
}
}
\references{
\url{https://ufal.mff.cuni.cz/udpipe}, \url{https://lindat.mff.cuni.cz/repository/xmlui/handle/11234/1-2364}, 
\url{http://universaldependencies.org/format.html}
}
\seealso{
\code{\link{udpipe_load_model}}
}
