% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udpipe_parse.R
\name{udpipe_load_model}
\alias{udpipe_load_model}
\title{Load an UDPipe model}
\usage{
udpipe_load_model(file)
}
\arguments{
\item{file}{full path to the model}
}
\value{
An object of class \code{udpipe_model} which is a list with 2 elements: file containing the file and model which is 
an Rcpp-generated pointer to the loaded model which can be used in \code{\link{udpipe_annotate}}
}
\description{
Load an UDPipe model
}
\details{
Pre-trained Universal Dependencies 2.0 models on all UD treebanks are made available at 
\url{https://ufal.mff.cuni.cz/udpipe}, namely at \url{https://lindat.mff.cuni.cz/repository/xmlui/handle/11234/1-2364}.
At the time of writing this consists of models made available on 50 languages, namely: 
ancient_greek, arabic, basque, belarusian, bulgarian, catalan, chinese, coptic, croatian, czech, danish, dutch, 
english, estonian, finnish, french, galician, german, gothic, greek, hebrew, hindi, hungarian, indonesian, irish, 
italian, japanese, kazakh, korean, latin, latvian, lithuanian, norwegian, old_church_slavonic, persian, polish, 
portuguese, romanian, russian, sanskrit, slovak, slovenian, spanish, swedish, tamil, turkish, ukrainian, 
urdu, uyghur, vietnamese. Mark that these models are made available under the CC BY-NC-SA 4.0 license.
}
\examples{
\dontrun{
## Download zipped folder with models + get one model
zips <- "https://lindat.mff.cuni.cz/repository/xmlui/bitstream/handle/11234/1-2364/udpipe-ud-2.0-170801.zip"
download.file(zips, "udpipe-ud-2.0-170801.zip")
unzip("udpipe-ud-2.0-170801.zip", list = TRUE)
unzip("udpipe-ud-2.0-170801.zip", files = "udpipe-ud-2.0-170801/dutch-ud-2.0-170801.udpipe", exdir = "dev")

f <- file.path(getwd(), "dev/udpipe-ud-2.0-170801/dutch-ud-2.0-170801.udpipe")
## Load model
ud_dutch <- udpipe_load_model(f)
}
}
\references{
\url{https://ufal.mff.cuni.cz/udpipe}, \url{https://lindat.mff.cuni.cz/repository/xmlui/handle/11234/1-2364}
}
\seealso{
\code{\link{udpipe_annotate}}
}
