% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unlist_tokens}
\alias{unlist_tokens}
\title{Create a data.frame from a list of tokens}
\usage{
unlist_tokens(x)
}
\arguments{
\item{x}{a list where the list elements are character vectors of tokens}
}
\value{
the data of \code{x} converted to a data.frame.
This data.frame has columns doc_id and token where the doc_id is taken from the list names of x
and token contains the data of \code{x}
}
\description{
Create a data.frame from a list of tokens.
}
\examples{
x <- setNames(c("some text here", "hi  there understand this?"), c("a", "b"))
x <- strsplit(x, split = " ")
x
unlist_tokens(x)
}
