% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udpipe_parse.R
\name{as.data.frame.udpipe_connlu}
\alias{as.data.frame.udpipe_connlu}
\title{Convert the result of udpipe_annotate to a tidy data frame}
\usage{
\method{as.data.frame}{udpipe_connlu}(x, ...)
}
\arguments{
\item{x}{an object of class \code{udpipe_connlu} as returned by \code{\link{udpipe_annotate}}}

\item{...}{currently not used}
}
\value{
a data.frame with columns doc_id, paragraph_id, sentence_id, sentence, 
token_id, token, lemma, upos, xpos, feats, head_token_id, dep_rel, deps, misc \cr

The columns paragraph_id, sentence_id are integers, the other fields
are character data in UTF-8 encoding. \cr

To get more information on these fields, visit \url{http://universaldependencies.org/format.html} 
or look at \code{\link{udpipe}}.
}
\description{
Convert the result of \code{\link{udpipe_annotate}} to a tidy data frame
}
\examples{
model    <- udpipe_download_model(language = "dutch-lassysmall")

if(!model$download_failed){

ud_dutch <- udpipe_load_model(model$file_model)
txt <- c("Ik ben de weg kwijt, kunt u me zeggen waar de Lange Wapper ligt? Jazeker meneer", 
         "Het gaat vooruit, het gaat verbazend goed vooruit")
x <- udpipe_annotate(ud_dutch, x = txt)
x <- as.data.frame(x)
head(x)

}

## cleanup for CRAN only - you probably want to keep your model if you have downloaded it
if(file.exists(model$file_model)) file.remove(model$file_model)
}
\seealso{
\code{\link{udpipe_annotate}}
}
