% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_flow.R
\name{dtm_align}
\alias{dtm_align}
\title{Reorder a Document-Term-Matrix alongside a vector or data.frame}
\usage{
dtm_align(x, y, FUN, ...)
}
\arguments{
\item{x}{a Document-Term-Matrix of class dgCMatrix (which can be an object returned by \code{\link{document_term_matrix}})}

\item{y}{either a vector or data.frame containing something to align with \code{x} (e.g. for predictive purposes).
\itemize{
\item{In case \code{y} is a vector, it should have names which are available in the rownames of \code{x}.}
\item{In case \code{y} is a vector, it's first column should contain identifiers which are available in the rownames of \code{x}.}
}}

\item{FUN}{a function to be applied on \code{x} before aligning it to \code{y}. See the examples}

\item{...}{further arguments passed on to FUN}
}
\value{
a list with elements \code{x} and \code{y} containing the document term matrix \code{x} in the same order as \code{y}.
\itemize{
\item{If in \code{y} a vector was passed, the returned \code{y} list element will be a vector}
\item{If in \code{y} a data.frame was passed with more than 2 columns, the returned \code{y} list element will be a data.frame, if it w}
\item{If in \code{y} a data.frame was passed with exactly 2 columns, the returned \code{y} list element will be a data.frame, if it w}
}
Only returns data of \code{x} with overlapping identifiers in \code{y}.
}
\description{
This utility function is useful to align a Document-Term-Matrix with 
information in a data.frame or a vector to predict, such that both the predictive information as well as the target 
is available in the same order. \cr
Matching is done based on the identifiers in the rownames of \code{x} and either the names of the vector \code{y} 
or the first column of the data.frame \code{y}.
}
\examples{
data(brussels_reviews)
data(brussels_listings)
x <- subset(brussels_reviews, language \%in\% "fr")
x <- strsplit.data.frame(x, term = "feedback", group = "listing_id")
x <- document_term_frequencies(x)
x <- document_term_matrix(x)
y <- brussels_listings$price
names(y) <- brussels_listings$listing_id

## align a matrix of predictors with a vector to predict
traindata <- dtm_align(x = x, y = y)
traindata <- dtm_align(x = x, y = y, FUN = function(dtm){
  dtm <- dtm_remove_lowfreq(dtm, minfreq = 5)
  dtm <- dtm_sample(dtm)
  dtm
})
head(names(y))
head(rownames(x))
head(names(traindata$y))
head(rownames(traindata$x))

## align a matrix of predictors with a data.frame
traindata <- dtm_align(x = x, y = brussels_listings[, c("listing_id", "price")])
traindata <- dtm_align(x = x, y = brussels_listings[, c("listing_id", "price", "room_type")])
head(traindata$y$listing_id)
head(rownames(traindata$x))
}
\seealso{
\code{\link{document_term_matrix}}
}
