% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_flow.R
\name{dtm_colsums}
\alias{dtm_colsums}
\alias{dtm_rowsums}
\title{Column sums and Row sums for document term matrices}
\usage{
dtm_colsums(dtm)

dtm_rowsums(dtm)
}
\arguments{
\item{dtm}{an object returned by \code{\link{document_term_matrix}}}
}
\value{
a vector of row/column sums with corresponding names
}
\description{
Column sums and Row sums for document term matrices
}
\examples{
x <- data.frame(
 doc_id = c(1, 1, 2, 3, 4), 
 term = c("A", "C", "Z", "X", "G"), 
 freq = c(1, 5, 7, 10, 0))
dtm <- document_term_matrix(x)
x <- dtm_colsums(dtm)
x
x <- dtm_rowsums(dtm)
head(x)
}
