% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/udpipe_parse.R
\name{udpipe_annotate}
\alias{udpipe_annotate}
\title{Tokenise, Tag and Dependency Parsing Annotation of raw text}
\usage{
udpipe_annotate(object, x, doc_id = paste("doc", seq_along(x), sep = ""),
  tokenizer = "tokenizer", tagger = c("default", "none"),
  parser = c("default", "none"), ...)
}
\arguments{
\item{object}{an object of class \code{udpipe_model} as returned by \code{\link{udpipe_load_model}}}

\item{x}{a character vector in UTF-8 encoding where each element of the character vector 
contains text which you like to tokenize, tag and perform dependency parsing.}

\item{doc_id}{an identifier of a document with the same length as \code{x}.}

\item{tokenizer}{a character string of length 1, which is either 'tokenizer' (default udpipe tokenisation)
or a character string with more complex tokenisation options 
as specified in \url{http://ufal.mff.cuni.cz/udpipe/users-manual} in which case \code{tokenizer} should be a character string where the options
are put after each other using the semicolon as separation.}

\item{tagger}{a character string of length 1, which is either 'default' (default udpipe POS tagging and lemmatisation)
or 'none' (no POS tagging and lemmatisation needed) or a character string with more complex tagging options 
as specified in \url{http://ufal.mff.cuni.cz/udpipe/users-manual} in which case \code{tagger} should be a character string where the options
are put after each other using the semicolon as separation.}

\item{parser}{a character string of length 1, which is either 'default' (default udpipe dependency parsing) or
'none' (no dependency parsing needed) or a character string with more complex parsing options 
as specified in \url{http://ufal.mff.cuni.cz/udpipe/users-manual} in which case \code{parser} should be a character string where the options
are put after each other using the semicolon as separation.}

\item{...}{currently not used}
}
\value{
a list with 3 elements
\itemize{
 \item{x: }{The \code{x} character vector with text.}
 \item{conllu: }{A character vector of length 1 containing the annotated result of the annotation flow in CONLL-U format.
 This format is explained at \url{http://universaldependencies.org/format.html}}
 \item{error: }{A vector with the same length of \code{x} containing possible errors when annotating \code{x}}
}
}
\description{
Tokenise, Tag and Dependency Parsing Annotation of raw text
}
\examples{
##
## Example on tagging Sanskrit
##  models available at https://github.com/jwijffels/udpipe.models.ud.2.0
##
url <- file.path("https://github.com/jwijffels/udpipe.models.ud.2.0/raw/master",
  "inst", "udpipe-ud-2.0-170801", 
  "sanskrit-ud-2.0-170801.udpipe")
f <- file.path(getwd(), "sanskrit-ud-2.0-170801.udpipe")
download.file(url = url, destfile = f)
ud_sanskrit <- udpipe_load_model(file = f)

txt <- 'ततः असौ प्राह क्षत्रियस्य तिस्रः भार्या धर्मम् भवन्ति तत् एषा कदाचिद् वैश्या
सुता भविष्यति तत् अनुरागः ममास्याम् ततः रथकारः तस्य निश्चयम् विज्ञायावदत् वयस्य किम् अ
धुना कर्तव्यम् कौलिकः आह किम् अहम् जानामि त्वयि मित्रे यत् अभिहितं मया ततः'
x <- udpipe_annotate(ud_sanskrit, x = txt)
as.data.frame(x)

##
## Example on Dutch
##  models available at https://github.com/jwijffels/udpipe.models.ud.2.0
##
\dontrun{
url <- file.path("https://github.com/jwijffels/udpipe.models.ud.2.0/raw/master",
  "inst", "udpipe-ud-2.0-170801", 
  "dutch-ud-2.0-170801.udpipe")
f <- file.path(getwd(), "dutch-ud-2.0-170801.udpipe")
download.file(url = url, destfile = f)
ud_dutch <- udpipe_load_model(file = f)

## Tokenise, Tag and Dependency Parsing Annotation. Output is in CONLL-U format.
txt <- c("Dus. Godvermehoeren met pus in alle puisten, 
  zei die schele van Van Bukburg en hij had nog gelijk ook. 
  Er was toen dat liedje van tietenkonttieten kont tieten kontkontkont, 
  maar dat hoefden we geenseens niet te zingen. 
  Je kunt zeggen wat je wil van al die gesluierde poezenpas maar d'r kwam wel 
  een vleeswarenwinkel onder te voorschijn van heb je me daar nou.
  
  En zo gaat het maar door.",
  "Wat die ransaap van een academici nou weer in z'n botte pan heb gehaald mag 
  Joost in m'n schoen gooien, maar feit staat boven water dat het een gore 
  vieze vuile ransaap is.")
x <- udpipe_annotate(ud_dutch, x = txt)
cat(x$conllu)
as.data.frame(x)

## Only tokenisation
x <- udpipe_annotate(ud_dutch, x = txt, tagger = "none", parser = "none")
as.data.frame(x)

## Only tokenisation and POS tagging + lemmatisation, no dependency parsing
x <- udpipe_annotate(ud_dutch, x = txt, tagger = "default", parser = "none")
as.data.frame(x)

## Only tokenisation and dependency parsing, no POS tagging nor lemmatisation
x <- udpipe_annotate(ud_dutch, x = txt, tagger = "none", parser = "default")
as.data.frame(x)

## Provide doc_id for joining and identification purpose
x <- udpipe_annotate(ud_dutch, x = txt, doc_id = c("id1", "feedbackabc"),
                     tagger = "none", parser = "none")
as.data.frame(x)
}
}
\references{
\url{https://ufal.mff.cuni.cz/udpipe}, \url{https://lindat.mff.cuni.cz/repository/xmlui/handle/11234/1-2364}, 
\url{http://universaldependencies.org/format.html}
}
\seealso{
\code{\link{udpipe_load_model}}, \code{\link{as.data.frame.udpipe_connlu}}
}
